#3516 fixing containerization build
