+++ .prettierignore
@@ -17,6 +17,7 @@
 
 # production
 /build
+/src/*/dist 
 
 # misc
 .DS_Store
