+++ src/backend/tests/test-utils.ts
@@ -770,57 +770,6 @@ export const createTestPart = async (
   return part;
 };
 
-export const CreatePartTag = async (organizationId: string, name: string, colorHexCode: string) => {
-  return await prisma.part_Tag.create({
-    data: {
-      name,
-      organization: {
-        connect: { organizationId }
-      },
-      colorHexCode,
-      dateCreated: new Date()
-    }
-  });
-};
-
-export const CreateCommonMistake = async (
-  title: string,
-  description: string,
-  starred: boolean,
-  user: User,
-  organizationId: string
-) => {
-  return await prisma.part_Review_Common_Mistake.create({
-    data: {
-      title,
-      description,
-      starred,
-      dateCreated: new Date(),
-      organization: {
-        connect: { organizationId }
-      },
-      userCreated: {
-        connect: { userId: user.userId }
-      }
-    }
-  });
-};
-
-export const CreatePartReviewFAQ = async (question: string, answer: string, organizationId: string, user: User) => {
-  return await prisma.frequentlyAskedQuestion.create({
-    data: {
-      question,
-      answer,
-      partReviewFaqOrg: {
-        connect: { organizationId }
-      },
-      userCreated: {
-        connect: { userId: user.userId }
-      }
-    }
-  });
-};
-
 export const createTestPartReview = async (
   partReviewId: string,
   fileIds: string[],
