+++ src/frontend/Dockerfile
@@ -10,7 +10,7 @@ COPY ./src/shared src/shared
 RUN yarn install;
 
 RUN yarn build:shared
-RUN yarn build:frontend:dev
+RUN export NODE_OPTIONS=--max_old_space_size=4096 && yarn build:frontend:dev
 
 FROM nginx:alpine
