+++ src/shared/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.build.json",
   "compilerOptions": {
-    "module": "esnext",
+    "module": "CommonJS",
     "target": "ES2019",
     "outDir": "./dist",
     "noEmit": false,
